+++ internal/paw/item_note.go
@@ -9,8 +9,6 @@ import (
 	"fyne.io/fyne/v2/data/binding"
 	"fyne.io/fyne/v2/layout"
 	"fyne.io/fyne/v2/widget"
-
-	"lucor.dev/paw/internal/icon"
 )
 
 func init() {
@@ -24,21 +22,24 @@ var _ Item = (*Note)(nil)
 var _ FyneObject = (*Note)(nil)
 
 type Note struct {
-	Value string
-	Metadata
+	Value     string `json:"value,omitempty"`
+	*Metadata `json:"metadata,omitempty"`
 }
 
 func NewNote() *Note {
 	return &Note{
-		Metadata: Metadata{
-			IconResource: icon.NoteOutlinedIconThemed,
-			Type:         NoteItemType,
+		Metadata: &Metadata{
+			Type: NoteItemType,
 		},
 	}
 }
 
 func (n *Note) Edit(ctx context.Context, w fyne.Window) (fyne.CanvasObject, Item) {
-	noteItem := *n
+	noteItem := &Note{}
+	*noteItem = *n
+	noteItem.Metadata = &Metadata{}
+	*noteItem.Metadata = *n.Metadata
+
 	titleEntry := widget.NewEntryWithData(binding.BindString(&noteItem.Metadata.Name))
 	titleEntry.Validator = nil
 	titleEntry.PlaceHolder = "Untitled note"
@@ -53,7 +54,7 @@ func (n *Note) Edit(ctx context.Context, w fyne.Window) (fyne.CanvasObject, Item
 	form.Add(labelWithStyle("Note"))
 	form.Add(noteEntry)
 
-	return form, &noteItem
+	return form, noteItem
 }
 
 func (n *Note) Show(ctx context.Context, w fyne.Window) fyne.CanvasObject {