+++ internal/paw/item_password.go
@@ -55,24 +55,23 @@ func (pm PasswordMode) String() string {
 }
 
 type Password struct {
-	Value  string
-	Format Format
-	Length int
-	Mode   PasswordMode
+	Value  string       `json:"value,omitempty"`
+	Format Format       `json:"format,omitempty"`
+	Length int          `json:"length,omitempty"`
+	Mode   PasswordMode `json:"mode,omitempty"`
 
-	Metadata
-	Note
+	*Metadata `json:"metadata,omitempty"`
+	*Note     `json:"note,omitempty"`
 
 	fpg FynePasswordGenerator
 }
 
 func NewPassword() *Password {
 	return &Password{
-		Metadata: Metadata{
-			IconResource: icon.PasswordOutlinedIconThemed,
-			Type:         PasswordItemType,
+		Metadata: &Metadata{
+			Type: PasswordItemType,
 		},
-		Note: Note{},
+		Note: &Note{},
 	}
 }
 
@@ -81,8 +80,13 @@ func (p *Password) SetPasswordGenerator(fpg FynePasswordGenerator) {
 }
 
 func (p *Password) Edit(ctx context.Context, w fyne.Window) (fyne.CanvasObject, Item) {
+	passwordItem := &Password{}
+	*passwordItem = *p
+	passwordItem.Metadata = &Metadata{}
+	*passwordItem.Metadata = *p.Metadata
+	passwordItem.Note = &Note{}
+	*passwordItem.Note = *p.Note
 
-	passwordItem := *p
 	passwordBind := binding.BindString(&passwordItem.Value)
 	titleEntry := widget.NewEntryWithData(binding.BindString(&passwordItem.Name))
 	titleEntry.Validator = nil
@@ -108,7 +112,7 @@ func (p *Password) Edit(ctx context.Context, w fyne.Window) (fyne.CanvasObject, 
 	})
 
 	passwordMakeButton := widget.NewButtonWithIcon("Generate", icon.KeyOutlinedIconThemed, func() {
-		p.fpg.ShowPasswordGenerator(passwordBind, &passwordItem, w)
+		p.fpg.ShowPasswordGenerator(passwordBind, passwordItem, w)
 	})
 
 	form := container.New(layout.NewFormLayout())
@@ -122,7 +126,7 @@ func (p *Password) Edit(ctx context.Context, w fyne.Window) (fyne.CanvasObject, 
 	form.Add(labelWithStyle("Note"))
 	form.Add(noteEntry)
 
-	return form, &passwordItem
+	return form, passwordItem
 }
 
 func (p *Password) Show(ctx context.Context, w fyne.Window) fyne.CanvasObject {